LayoutTests/platform/chromium/fast/dom/prototype-inheritance-expected.txt

Make sure prototypes are set up using the window a property came from, instead of the lexical global object.

On success, you will see no "FAIL" messages, followed by "TEST COMPLETE".


FAIL inner.applicationCache.isInner should be true. Was false.
FAIL inner.applicationCache.constructor.isInner should be true. Was false.
FAIL inner.blur.isInner should be true. Was false.
FAIL inner.blur.constructor.isInner should be true. Was false.
FAIL inner.clientInformation.isInner should be true. Was false.
FAIL inner.clientInformation.constructor.isInner should be true. Was false.
FAIL inner.close.isInner should be true. Was false.
FAIL inner.close.constructor.isInner should be true. Was false.
FAIL inner.console.isInner should be true. Was false.
FAIL inner.console.constructor.isInner should be true. Was false.
FAIL inner.crypto.isInner should be true. Was false.
FAIL inner.crypto.constructor.isInner should be true. Was false.
FAIL inner.focus.isInner should be true. Was false.
FAIL inner.focus.constructor.isInner should be true. Was false.
FAIL inner.frameElement.isInner should be true. Was false.
FAIL inner.frameElement.constructor.isInner should be true. Was false.
FAIL inner.localStorage.isInner should be true. Was false.
FAIL inner.localStorage.constructor.isInner should be true. Was false.
FAIL inner.locationbar.isInner should be true. Was false.
FAIL inner.locationbar.constructor.isInner should be true. Was false.
FAIL inner.menubar.isInner should be true. Was false.
FAIL inner.menubar.constructor.isInner should be true. Was false.
FAIL inner.navigator.isInner should be true. Was false.
FAIL inner.navigator.constructor.isInner should be true. Was false.
FAIL inner.personalbar.isInner should be true. Was false.
FAIL inner.personalbar.constructor.isInner should be true. Was false.
FAIL inner.postMessage.isInner should be true. Was false.
FAIL inner.postMessage.constructor.isInner should be true. Was false.
FAIL inner.screen.isInner should be true. Was false.
FAIL inner.screen.constructor.isInner should be true. Was false.
FAIL inner.scrollbars.isInner should be true. Was false.
FAIL inner.scrollbars.constructor.isInner should be true. Was false.
FAIL inner.sessionStorage.isInner should be true. Was false.
FAIL inner.sessionStorage.constructor.isInner should be true. Was false.
FAIL inner.statusbar.isInner should be true. Was false.
FAIL inner.statusbar.constructor.isInner should be true. Was false.
FAIL inner.styleMedia.isInner should be true. Was false.
FAIL inner.styleMedia.constructor.isInner should be true. Was false.
FAIL inner.toolbar.isInner should be true. Was false.
FAIL inner.toolbar.constructor.isInner should be true. Was false.
FAIL inner.webkitIndexedDB.isInner should be true. Was false.
FAIL inner.webkitIndexedDB.constructor.isInner should be true. Was false.
FAIL inner.webkitPostMessage.isInner should be true. Was false.
FAIL inner.webkitPostMessage.constructor.isInner should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
